refactor(SPE-281): split attrition reset and share derived-state recompute
Agent-Logs-Url: https://github.com/JamesJedi420/containment-protocol/sessions/31f43fae-bba5-40c7-952d-7d3324cf421b
