+++ src/app/store/gameStore.ts
@@ -80,7 +80,7 @@ import { createSeededRng, normalizeSeed } from '../../domain/math'
 import type { EquipmentSlotKind } from '../../domain/equipment'
 import { discardPartyCard, drawPartyCards, playPartyCard } from '../../domain/partyCards/engine'
 import { createStartingState } from '../../data/startingState'
-import { applyChapterBreakAttritionReset } from '../../domain/agent/attritionContinuity'
+import { applyChapterBreakAttritionReset } from '../../domain/agent/attritionReset'
 import { advanceWeek } from '../../domain/sim/advanceWeek'
 import { assignTeam, launchMajorIncident, unassignTeam } from '../../domain/sim/assign'
 import { queueFabrication } from '../../domain/sim/production'
